LPS-66050 Simplify
