+++ portal-kernel/src/com/liferay/portal/kernel/service/persistence/impl/BasePersistenceImpl.java
@@ -14,6 +14,7 @@
 
 package com.liferay.portal.kernel.service.persistence.impl;
 
+import com.liferay.expando.kernel.model.ExpandoBridge;
 import com.liferay.portal.kernel.configuration.Filter;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
@@ -548,8 +549,7 @@ public class BasePersistenceImpl<T extends BaseModel<T>>
 	private SessionFactory _sessionFactory;
 
 	private static class NullModel
-		extends BaseModelAdapter<NullModel>
-		implements CacheModel<NullModel>, MVCCModel {
+		implements BaseModel<NullModel>, CacheModel<NullModel>, MVCCModel {
 
 		@Override
 		public Object clone() {
@@ -557,8 +557,63 @@ public class BasePersistenceImpl<T extends BaseModel<T>>
 		}
 
 		@Override
+		public int compareTo(NullModel nullModel) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public ExpandoBridge getExpandoBridge() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public Map<String, Object> getModelAttributes() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public Class<?> getModelClass() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public String getModelClassName() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
 		public long getMvccVersion() {
 			return -1;
+		}
+
+		@Override
+		public Serializable getPrimaryKeyObj() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public boolean isCachedModel() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public boolean isEntityCacheEnabled() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public boolean isEscapedModel() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public boolean isFinderCacheEnabled() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public boolean isNew() {
+			throw new UnsupportedOperationException();
 		}
 
 		@Override
@@ -570,7 +625,37 @@ public class BasePersistenceImpl<T extends BaseModel<T>>
 		}
 
 		@Override
+		public void setExpandoBridgeAttributes(BaseModel<?> baseModel) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public void setExpandoBridgeAttributes(ExpandoBridge expandoBridge) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public void setExpandoBridgeAttributes(ServiceContext serviceContext) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public void setModelAttributes(Map<String, Object> attributes) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
 		public void setMvccVersion(long mvccVersion) {
+		}
+
+		@Override
+		public void setNew(boolean n) {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public void setPrimaryKeyObj(Serializable primaryKeyObj) {
+			throw new UnsupportedOperationException();
 		}
 
 		@Override
@@ -581,6 +666,21 @@ public class BasePersistenceImpl<T extends BaseModel<T>>
 		@Override
 		public NullModel toEntityModel() {
 			return this;
+		}
+
+		@Override
+		public NullModel toEscapedModel() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public NullModel toUnescapedModel() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public String toXmlString() {
+			throw new UnsupportedOperationException();
 		}
 
 	}
